Add Docker stage to query PaperMC Download API
As noted in #2, Versions of Velocity > 3.1.0 are only accessible via a Build number which has to be queried via API.  In order to facilitate this, a script has been written that queries the three API endpoints needed to get a download path:

- Versions API
- Builds API
- Downloads API

As these are JSON APIs and the openjdk11:alpine-slim image contains no tooling to do JSON parsing, a builder stage was added to the Dockerfile to determine the download path and download the JAR.  This JAR is then copied into the final build stage to keep the Docker image small.

This could have been solved by pulling in jq via apk into the final image and running a shell script but that would have been much harder to read and more likely to bust a layer cache due to the update.